Separated HTML-generating from JSON-generating controllers.
